Create Add-user_to_a_AD group
Import Active Directory Module: Ensures the Active Directory module is available for use.
Prompt for AD Group Name: Uses Read-Host to get the group name from the user.
Get the Group Object: Uses Get-ADGroup to find the group. If the group doesn't exist, it will notify the user.
Prompt for User to Add: Uses Read-Host to get the username to be added to the group.
Get the User Object: Uses Get-ADUser to find the user. If the user doesn't exist, it will notify the user.
Add User to Group: Uses Add-ADGroupMember to add the specified user to the group.